# docs/swagger.yaml
basePath: /
definitions:
  response.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  users.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  users.User:
    properties:
      address:
        type: string
      avatar:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_blocked:
        type: boolean
      is_deleted:
        type: boolean
      is_suspended:
        type: boolean
      is_verified:
        type: boolean
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
host: ${VERCEL_URL}
info:
  contact:
    email: your-email@example.com
    name: Your Name
    url: https://your-website.com
  description: This is a simple CMS API.
  title: CMS Project API
  version: "1.0"
paths:
  /api/nav_links:
    get:
      description: Retrieve a list of all navigation links (menus)
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get navigation links
      tags:
      - NavLink
  /api/users:
    get:
      description: Retrieve a list of all users from the database
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get all users
      tags:
      - User
  /api/users/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/users.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: User login
      tags:
      - User
  /api/users/register:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/users.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Creates a new user
      tags:
      - User
schemes:
- https
swagger: "2.0"